Moving exercises around

+++ src/com/interview/sde/hackerrank/java/advanced/IsBST.java
@@ -1,13 +1,11 @@
 package com.interview.sde.hackerrank.java.advanced;
 
-import java.util.*;
-
 public class IsBST {
     boolean checkBST(InsertBST.Node root) {
-        return checkInternal(root, Integer.MIN_VALUE, Integer.MAX_VALUE);
+        return checkInternal(root, Long.MIN_VALUE, Long.MAX_VALUE);
     }
 
-    boolean checkInternal(InsertBST.Node root, int min, int max) {
+    boolean checkInternal(InsertBST.Node root, long min, long max) {
         return root == null || (root.data > min && root.data < max && checkInternal(root.left, min, root.data) && checkInternal(root.right, root.data, max));
     }
 }
